Reorganizar importaciones para utilizar rutas compartidas en componentes de clientes
