+++ resources/js/components/clients/clientes-details.tsx
@@ -8,8 +8,8 @@ import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table"
-import ConfirmDeleteDialog from "../confirm-dialog"
-import { NoData } from "../no-data"
+import ConfirmDeleteDialog from "../shared/confirm-dialog"
+import { NoData } from "../shared/no-data"
 interface DetallesClienteProps {
     cliente: Cliente
     onClose: () => void
