+++ resources/js/components/clients/clientes-tabla.tsx
@@ -1,16 +1,16 @@
 import { useState } from "react"
 import { Search, Info, Edit, Plus, Trash2, PlusCircle } from "lucide-react"
 import DetallesCliente from "./clientes-details"
-import { NoData } from "../no-data"
+import { NoData } from "../shared/no-data"
 import type { Cliente, Cita } from "@/types/clients"
 
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-import ConfirmDeleteDialog from "../confirm-dialog"
+import ConfirmDeleteDialog from "../shared/confirm-dialog"
 import { toast } from "sonner"
-import { AppointmentDialog } from "../calendar/appointment-dialog"
+import { AppointmentDialog } from "../appointments/appointment-dialog"
 import { router } from "@inertiajs/react"
 import { category } from "@/types/services"
 import { format } from "date-fns"
